Wydzielenie części kodu z Parsera do Design Extractora. Dodanie możliwości parsowania paru procedur.
