+++ src/main/java/aitsi/m3spin/spafrontend/parser/Parser.java
@@ -13,136 +13,23 @@ import java.util.List;
 
 public class Parser {
     @Getter
-    private final Pkb pkb;
     private final CodeScanner codeScanner;
 
-    public Parser(List<String> code, Pkb pkb) {
+    public Parser(List<String> code) {
         this.codeScanner = new CodeScanner(code);
-        this.pkb = pkb;
     }
 
     public List<Procedure> parse() throws SimpleParserException {
         System.out.println("Parsing SIMPLE code...");
-        Procedure rootProc = parseProcedure();
+
+        List<Procedure> procedures = new ArrayList<>();
+        while (codeScanner.isEndOfFile()){
+            procedures.add(parseProcedure());
+            this.codeScanner.skipWhitespaces();
+        }
+
         System.out.println("Parsing completed.");
-        return Collections.singletonList(rootProc);
-    }
-
-    /*
-     * Wzór metody fillPkb(tnode):
-     * 1. insertVar(tnode) / insertProc(tnode)
-     * 2. setChild(tnode, tnodeChild) / setSibling(tnode, tnodeSibling)
-     * 3. Wywołaj fillPkb() dla tego dziecka/rodzeństwa
-     *
-     *  todo w przyszłuch iteracjach: Calls
-     * */
-    public void fillPkb(List<Procedure> procedures) throws UnknownStatementType {
-        System.out.println("Filling PKB with data...");
-        Procedure rootProc = procedures.get(0);
-        rootProc = (Procedure) pkb.getAst().setRoot(rootProc);
-        fillPkb(rootProc);
-        System.out.println("Filling PKB with data completed.");
-    }
-
-    private void fillPkb(Procedure procedure) throws UnknownStatementType {
-        pkb.getProcTable().insertProc(procedure.getName());
-        StatementList stmtList = (StatementList) pkb.getAst().setChild(procedure, procedure.getStatementList());
-        RelationshipsInfo relationshipsInfo = fillPkb(stmtList);
-
-        relationshipsInfo.getModifiedVariables()
-                .forEach(modifiedVar -> pkb.getModifiesInterface().setModifies(procedure, modifiedVar));
-
-        relationshipsInfo.getUsedVariables()
-                .forEach(usedVar -> pkb.getUsesInterface().setUses(procedure, usedVar));
-    }
-
-    private RelationshipsInfo fillPkb(StatementList stmtList) throws UnknownStatementType {
-        Statement firstStmt = (Statement) pkb.getAst().setChild(stmtList, stmtList.getStatements().get(0));
-        RelationshipsInfo stmtListRelationshipsInfo = fillPkb(firstStmt);
-
-        Statement currentStmt = firstStmt;
-
-        for (int i = 1; i < stmtList.getStatements().size(); i++) {
-            pkb.getFollowsInterface().setFollows(currentStmt, stmtList.getStatements().get(i));
-            currentStmt = (Statement) pkb.getAst().setSibling(currentStmt, stmtList.getStatements().get(i));
-            stmtListRelationshipsInfo = RelationshipsInfo.merge(stmtListRelationshipsInfo, fillPkb(currentStmt));
-        }
-
-        return stmtListRelationshipsInfo;
-    }
-
-    private RelationshipsInfo fillPkb(Statement stmt) throws UnknownStatementType {
-        if (stmt instanceof Assignment) {
-            return fillPkb((Assignment) stmt);
-        } else if (stmt instanceof While) {
-            return fillPkb((While) stmt);
-        } else if (stmt instanceof If) {
-            return RelationshipsInfo.emptyInfo();
-            //todo w przyszłych iteracjach
-        } else if (stmt instanceof Call) {
-            return RelationshipsInfo.emptyInfo();
-            //todo w przyszłych iteracjach
-        } else {
-            throw new UnknownStatementType();
-        }
-    }
-
-    private RelationshipsInfo fillPkb(While whileStmt) throws UnknownStatementType {
-        Variable conditionVar = (Variable) pkb.getAst().setChild(whileStmt, whileStmt.getConditionVar());
-        RelationshipsInfo relationshipsInfo = fillPkb(conditionVar, whileStmt.getStmtList());
-
-        relationshipsInfo.getModifiedVariables()
-                .forEach(modifiedVar -> pkb.getModifiesInterface().setModifies(whileStmt, modifiedVar));
-
-        pkb.getUsesInterface().setUses(whileStmt, conditionVar);
-        relationshipsInfo.getUsedVariables()
-                .forEach(usedVar -> pkb.getUsesInterface().setUses(whileStmt, usedVar));
-
-        whileStmt.getStmtList().getStatements()
-                .forEach(statement -> pkb.getParentInterface().setParent(whileStmt, statement));
-
-        return relationshipsInfo;
-    }
-
-    private RelationshipsInfo fillPkb(Assignment assignment) {
-        VariableImpl variable = (VariableImpl) pkb.getAst().setChild(assignment, assignment.getVariable());
-
-        RelationshipsInfo relationshipsInfo = new RelationshipsInfo();
-
-        relationshipsInfo.addModifiedVar(variable);
-        pkb.getModifiesInterface().setModifies(assignment, variable);
-
-        RelationshipsInfo.merge(relationshipsInfo, fillPkb(variable, assignment.getExpression()));
-        relationshipsInfo.getUsedVariables()
-                .forEach(usedVar -> pkb.getUsesInterface().setUses(assignment, usedVar));
-        return relationshipsInfo;
-    }
-
-    private RelationshipsInfo fillPkb(Variable variable, Expression expression) {
-        pkb.getVarTable().insertVar(variable.getName());
-        expression = (Expression) pkb.getAst().setSibling(variable, expression);
-        return fillPkb(expression);
-    }
-
-    private RelationshipsInfo fillPkb(Variable variable, StatementList stmtList) throws UnknownStatementType {
-        pkb.getVarTable().insertVar(variable.getName());
-        stmtList = (StatementList) pkb.getAst().setSibling(variable, stmtList);
-        return fillPkb(stmtList);
-    }
-
-    private RelationshipsInfo fillPkb(Expression expression) {
-        if (expression.getExpression() != null) {
-            RelationshipsInfo relationshipsInfo = new RelationshipsInfo();
-            if (expression.getFactor() instanceof Variable) {
-                Variable variable = (Variable) expression.getFactor();
-                pkb.getVarTable().insertVar(variable.getName());
-                relationshipsInfo.addUsedVar(variable);
-            }
-            Factor factor = (Factor) pkb.getAst().setChild(expression, expression.getFactor());
-
-            Expression nestedExpression = (Expression) pkb.getAst().setSibling(factor, expression.getExpression());
-            return RelationshipsInfo.merge(relationshipsInfo, fillPkb(nestedExpression));
-        } else return RelationshipsInfo.emptyInfo();
+        return procedures;
     }
 
     private Procedure parseProcedure() throws SimpleParserException {
@@ -154,7 +41,7 @@ public class Parser {
         Procedure procedure = new ProcedureImpl(procName, parseStmtList());
 
         //Whole endingBrace
-        parseChar('}', false);
+        parseChar('}', true);
 
         return procedure;
     }
